+++ changelog.md
@@ -3,6 +3,7 @@
 #v1.3.6
 * Tuned the USB period size for SDM845 devices (2500 usec to 2250 usec)
 * Tuned the USB period size for other devices (to 2250 usec)
+* Added checking incompatible Magisk variants
 
 # v1.3.5
 * Changed the re-sampling parameters for Galaxy S4 to the general purpose ones (optimized for 3.5mm jack; not USB DAC's)